Change positions and sizes from INT to FLOAT

+++ RocketStats/RocketStats.h
@@ -119,10 +119,10 @@ struct StatEventParams {
 };
 
 struct Vector2D {
-	int x = 0;
-	int y = 0;
-	int width = 0;
-	int height = 0;
+	float x = 0;
+	float y = 0;
+	float width = 0;
+	float height = 0;
 };
